backend/requirements.txt
fastapi>=0.136.0
uvicorn>=0.46.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.27
pydantic>=2.13.0
sqlalchemy>=2.0.0
akshare>=1.14.0
pandas>=2.0.0
aiofiles>=25.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
Pillow>=10.0.0
httpx>=0.27.0
